+++ crates/galfus-vm/src/runtime/tests/arithmetic_and_control.rs
@@ -18,7 +18,7 @@ fn test_basic_arithmetic() {
         },
         Instruction::Ret { src: Reg(3) },
     ];
-    let image = create_test_image(instrs, vec![Constant::Int(10), Constant::Int(20)]);
+    let image = create_test_image(instrs, vec![Constant::Int64(10), Constant::Int64(20)]);
     let mut vm = VirtualMachine::new(image);
     let res = vm.run_function(FuncIdx(0), vec![]).unwrap();
     assert_eq!(res, Value::Int64(30));
@@ -62,7 +62,7 @@ fn test_sub_mul_div_rem_pow() {
         }, // 3^4 = 81
         Instruction::Ret { src: Reg(7) },
     ];
-    let image = create_test_image(instrs, vec![Constant::Int(15), Constant::Int(4)]);
+    let image = create_test_image(instrs, vec![Constant::Int64(15), Constant::Int64(4)]);
     let mut vm = VirtualMachine::new(image);
     let res = vm.run_function(FuncIdx(0), vec![]).unwrap();
     assert_eq!(res, Value::Int64(81));
@@ -81,7 +81,7 @@ fn test_neg() {
         }, // -5
         Instruction::Ret { src: Reg(2) },
     ];
-    let image = create_test_image(instrs, vec![Constant::Int(5)]);
+    let image = create_test_image(instrs, vec![Constant::Int64(5)]);
     let mut vm = VirtualMachine::new(image);
     let res = vm.run_function(FuncIdx(0), vec![]).unwrap();
     assert_eq!(res, Value::Int64(-5));
@@ -119,7 +119,7 @@ fn test_bitnot() {
         }, // !5
         Instruction::Ret { src: Reg(2) },
     ];
-    let image = create_test_image(instrs, vec![Constant::Int(5)]);
+    let image = create_test_image(instrs, vec![Constant::Int64(5)]);
     let mut vm = VirtualMachine::new(image);
     let res = vm.run_function(FuncIdx(0), vec![]).unwrap();
     assert_eq!(res, Value::Int64(!5));
@@ -163,7 +163,7 @@ fn test_shl_shr_and_or_xor() {
         }, // 10 ^ 2 = 8
         Instruction::Ret { src: Reg(7) },
     ];
-    let image = create_test_image(instrs, vec![Constant::Int(8), Constant::Int(2)]);
+    let image = create_test_image(instrs, vec![Constant::Int64(8), Constant::Int64(2)]);
     let mut vm = VirtualMachine::new(image);
     let res = vm.run_function(FuncIdx(0), vec![]).unwrap();
     assert_eq!(res, Value::Int64(8));
@@ -187,7 +187,7 @@ fn test_comparison_lt() {
         }, // true
         Instruction::Ret { src: Reg(3) },
     ];
-    let image = create_test_image(instrs, vec![Constant::Int(100), Constant::Int(200)]);
+    let image = create_test_image(instrs, vec![Constant::Int64(100), Constant::Int64(200)]);
     let mut vm = VirtualMachine::new(image);
     let res = vm.run_function(FuncIdx(0), vec![]).unwrap();
     assert_eq!(res, Value::Bool(true));
@@ -211,7 +211,7 @@ fn test_comparison_le() {
         }, // false
         Instruction::Ret { src: Reg(3) },
     ];
-    let image = create_test_image(instrs, vec![Constant::Int(100), Constant::Int(200)]);
+    let image = create_test_image(instrs, vec![Constant::Int64(100), Constant::Int64(200)]);
     let mut vm = VirtualMachine::new(image);
     let res = vm.run_function(FuncIdx(0), vec![]).unwrap();
     assert_eq!(res, Value::Bool(false));
@@ -232,7 +232,7 @@ fn test_fallback() {
         }, // 100
         Instruction::Ret { src: Reg(3) },
     ];
-    let image = create_test_image(instrs, vec![Constant::Int(100)]);
+    let image = create_test_image(instrs, vec![Constant::Int64(100)]);
     let mut vm = VirtualMachine::new(image);
     let res = vm.run_function(FuncIdx(0), vec![]).unwrap();
     assert_eq!(res, Value::Int64(100));
@@ -265,8 +265,8 @@ fn test_control_flow_jumps() {
         instrs,
         vec![
             Constant::Bool(false),
-            Constant::Int(999),
-            Constant::Int(888),
+            Constant::Int64(999),
+            Constant::Int64(888),
         ],
     );
     let mut vm = VirtualMachine::new(image);
@@ -316,7 +316,7 @@ fn test_nested_calls_return_to_explicit_destinations() {
     let image = ModuleImage {
         name: "test".to_string(),
         constants: ConstantPool {
-            constants: vec![Constant::Int(1), Constant::Int(2)],
+            constants: vec![Constant::Int64(1), Constant::Int64(2)],
         },
         functions: vec![
             ImageFunction {
@@ -363,7 +363,7 @@ fn test_dynamic_call_returns_to_destination() {
     let image = ModuleImage {
         name: "test".to_string(),
         constants: ConstantPool {
-            constants: vec![Constant::Function(FuncIdx(1)), Constant::Int(7)],
+            constants: vec![Constant::Function(FuncIdx(1)), Constant::Int64(7)],
         },
         functions: vec![
             ImageFunction {
